add method in mapper

+++ discount-service/src/main/java/com/service/discount/mappers/Convert.java
@@ -21,7 +21,7 @@ public class Convert {
   private DiscountTypeRepository discountTypeRepository;
 
   @Named("idToDiscountType")
-  public DiscountType idToDiscountType(Long discountType) {
+  public DiscountType idToDiscountType(Integer discountType) {
     return discountTypeRepository.findById(discountType)
         .orElseThrow(() -> new RuntimeException("Discount type not found"));
   }
